tron: replace gRPC client with HTTP client (#254)
* tron/relayer/gotron-sdk: remove all gRPC, protobuf, and cmd code

* tron/relayer/gotron-sdk/pkg/abi/abi.go: remove proto/core referenced functions

* tron/relayer/gotron-sdk/pkg: remove account, keys, keystore, ledger, mnemonic, store packages

* tron/relayer/gotron-sdk/go.mod: update

* tron/relayer/sdk: update interfaces and combined client for HTTP API

* tron/relayer/mocks: update FullNodeClient mock

* tron/integration-tests/go.mod: update deps, fix broken grafana build

* tron/relayer/ocr2/config_digester.go: remove debug log

* tron/relayer/ocr2/contract_transmitter.go: update Enqueue call

* tron/relayer/plugin/relayer.go: update client creation to FullNodeClient

* tron/relayer/reader/reader.go: use common.JSONABI

* tron/relayer/testutils/tron_node.go: add HTTP URLs, remove gRPC URLs

* tron/relayer/testutils/txm.go: update to use FullNodeClient

* tron/relayer/testutils/utils.go: dont use gotron-sdk keystore

* tron/relayer/txm: use int32 for energy units to match HTTP API

* tron/relayer/utils.go: remove GetFunctionSignature, event hash returns a hex string

* tron/relayer/utils_test.go: update tests for common.JSONABI

* tron/relayer/txm: use full node client, address.Address params instead of strings

* tron/relayer/monitor/balance.go: use full node client

* tron/integration-tests: use http clients

* tron/integration-tests/ocr2_local_test.go: fix transmitter addresses

* tron/scripts/core.sh: dont hide errors when creating database fails

* tron/relayer/reader/reader.go: fix matching contract addresses

* tron/relayer/reader/reader_test.go: fix tests

* tron/relayer/ocr2/ocr2_reader_test.go: fix tests

* tron/relayer: remove unused testutils DeployContract

* tron: rename deploy contract helper to SignAndDeployContract

* tron/relayer/testutils/txm.go: read the txid from broadcast response instead

+++ relayer/txm/energy_test.go
@@ -17,7 +17,7 @@ func TestParseLatestEnergyPrice(t *testing.T) {
 	testCases := []struct {
 		name            string
 		energyPricesStr string
-		expectedPrice   int64
+		expectedPrice   int32
 		expectedErrMsg  string
 	}{
 		{
@@ -55,9 +55,9 @@ func TestParseLatestEnergyPrice(t *testing.T) {
 
 func TestCalculatePaddedFeeLimit(t *testing.T) {
 	tests := []struct {
-		feeLimit int64
-		attempt  uint64
-		expected int64
+		feeLimit int32
+		attempt  uint32
+		expected int32
 	}{
 		{feeLimit: 1000, attempt: 0, expected: 1000},
 		{feeLimit: 1000, attempt: 1, expected: 1500},
